refactor: split transaction import API workflow
