+++ frontend/src/api/transactionImports/hooks.ts
@@ -1,0 +1,16 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { invalidateAppData } from '@/api/cacheInvalidation';
+import { importTransactionsInBatches } from '@/api/transactionImports/batching';
+
+/**
+ * Provides the mutation boundary for uploading prepared transaction import payloads
+ */
+export function useImportTransactions() {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: importTransactionsInBatches,
+    onSuccess: () => {
+      invalidateAppData(queryClient);
+    },
+  });
+}
